+++ src/components/userRegistration/UserRegistration.scss
@@ -1,6 +1,6 @@
 .userRegistrationWrapper {
-  width: 20%;
-  height: 30%;
+  width: 350px;
+  height: 350px;
   position: absolute;
   top: 50%;
   left: 50%;
@@ -8,4 +8,11 @@
   background: white;
   border-radius: 50%;
   padding: 3%;
+
+  .formContainer {
+    width: 75%;
+    height: 75%;
+    margin: 3% auto;
+    
+  }
 }
