+++ src/views/views.scss
@@ -14,6 +14,8 @@
   background: black;
   color: white;
   text-align: center;
+  display: flex;
+  flex-direction: column;
 }
 
 .landingTitle {
